+++ CHANGELOG.md
@@ -7,15 +7,19 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ### Oneview Features supported
 - Enclosure
+- Enclosure Group
 - FC Network
 - FCoE Network
 - Interconnect
 - Interconnect type
+- Logical Enclosure
 - Logical interconnect
+- Logical Interconnect Group
 - Network Set
 - Server hardware
 - Server hardware type
 - Server Profile
+- Server Profile Template
 - Storage Pool
 - Storage System
 - Storage Volume
@@ -53,6 +57,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - Scope
 - Server hardware type
 - Server Profile
+- Server Profile Template
 - Storage pool
 - Storage system
 - Storage volume